+++ src/shared/index.ts
@@ -1,5 +1,7 @@
 export { cn } from './lib/utils';
 
+export { HEADERS_TEXT } from './lib/locales';
+
 export { ROUTES } from './config/routes';
 
 export { useTheme } from './hooks/use-theme';
